Write my own assign, replaces clone
