+++ lib/Element.js
@@ -4,7 +4,7 @@ var querySelectorAll = require('query-selector')
 var camelCase = require('./utils/camelCase')
 var isString = require('./utils/isString')
 var isUndefined = require('./utils/isUndefined')
-var clone = require('./utils/clone')
+var assign = require('./utils/assign')
 var mapValues = require('./utils/mapValues')
 var styleCamelCase = require('./utils/styleCamelCase')
 
@@ -231,8 +231,8 @@ Element.prototype.getElementById = function (id) {
 
 Element.prototype.toReact = function (index) {
   index = index || 0
-  var props = clone(this.props)
-  props.style = clone(props.style)
+  var props = assign({}, this.props)
+  props.style = assign({}, props.style)
 
   var originalElement = this
 
@@ -248,7 +248,7 @@ Element.prototype.toReact = function (index) {
   delete props.style.getProperty
   delete props.style.removeProperty
 
-  Object.assign(props, mapValues(this.eventListeners, function (listeners) {
+  assign(props, mapValues(this.eventListeners, function (listeners) {
     return function (syntheticEvent) {
       var event
