--- lib/utils/clone.js
@@ -1,9 +1,0 @@
-var mapValues = require('./mapValues')
-
-function clone (value) {
-  return mapValues(value, function (n) {
-    return n
-  })
-}
-
-module.exports = clone
